% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tdmurl.R
\name{as_tdmurl}
\alias{as_tdmurl}
\alias{as_tdmurl.character}
\alias{as_tdmurl.tdmurl}
\title{Coerce a url to a tdmurl with a specific type}
\usage{
as_tdmurl(url, type, doi)

\method{as_tdmurl}{tdmurl}(url, type, doi)

\method{as_tdmurl}{character}(url, type, doi = NULL)
}
\arguments{
\item{url}{A URL.}

\item{type}{A document type, one of xml, pdf, or plain}

\item{doi}{A DOI, optional, defaults to \code{NULL}}
}
\description{
Coerce a url to a tdmurl with a specific type
}
\examples{
\dontrun{
as_tdmurl("http://downloads.hindawi.com/journals/bmri/2014/201717.xml",
   "xml")
as_tdmurl("http://downloads.hindawi.com/journals/bmri/2014/201717.pdf",
   "pdf")
out <- as_tdmurl("http://downloads.hindawi.com/journals/bmri/2014/201717.pdf",
   "pdf", "10.1155/2014/201717")
attributes(out)
}
}

