% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_extract.R
\name{crm_extract}
\alias{crm_extract}
\title{Extract text from a single pdf document.}
\usage{
crm_extract(path)
}
\arguments{
\item{path}{Path to a file}
}
\value{
An object of class xpdf_char
}
\description{
Extract text from a single pdf document.
}
\examples{
\dontrun{
path <- system.file("examples", "MairChamberlain2014RJournal.pdf",
   package = "crminer")
(res_xpdf <- crm_extract(path))
res_xpdf$meta
res_xpdf$data
}
}

