% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_text.R
\name{crm_text}
\alias{crm_text}
\alias{crm_plain}
\alias{crm_xml}
\alias{crm_pdf}
\title{Get full text from a DOI}
\usage{
crm_text(url, type = "xml", path = cr_cache_path(), overwrite = TRUE,
  read = TRUE, cache = TRUE, overwrite_unspecified = FALSE, ...)

crm_plain(url, path = cr_cache_path(), overwrite = TRUE, read = TRUE,
  overwrite_unspecified = FALSE, ...)

crm_xml(url, path = cr_cache_path(), overwrite = TRUE, read = TRUE,
  overwrite_unspecified = FALSE, ...)

crm_pdf(url, path = cr_cache_path(), overwrite = TRUE, read = TRUE,
  cache = FALSE, overwrite_unspecified = FALSE, ...)
}
\arguments{
\item{url}{(character) A URL.}

\item{type}{(character) One of 'xml' (default), 'html', 'plain', 'pdf',
'unspecified', or 'all'}

\item{path}{(character) Path to store pdfs in. By default we use
\code{paste0(rappdirs::user_cache_dir(), "/crminer")}, but you can
set this directory to something different. Ignored unless getting
pdf}

\item{overwrite}{(logical) Overwrite file if it exists already?
Default: \code{TRUE}}

\item{read}{(logical) If reading a pdf, this toggles whether we extract
text from the pdf or simply download. If \code{TRUE}, you get the text from
the pdf back. If \code{FALSE}, you only get back the metadata.
Default: \code{TRUE}}

\item{cache}{(logical) Use cached files or not. All files are written to
your machine locally, so this doesn't affect that. This only states whether
you want to use cached version so that you don't have to download the file
again. The steps of extracting and reading into R still have to be performed
when \code{cache=TRUE}. Default: \code{TRUE}}

\item{overwrite_unspecified}{(logical) Sometimes the crossref API returns
mime type 'unspecified' for the full text links (for some Wiley dois
for example). This parameter overrides the mime type to be \code{type}.}

\item{...}{Named parameters passed on to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\description{
Get full text from a DOI
}
\details{
Note that \code{\link[=crm_text]{crm_text()}}, \code{\link[=crm_pdf]{crm_pdf()}}, \code{\link[=crm_xml]{crm_xml()}}, \code{\link[=crm_plain]{crm_plain()}}
are not vectorized.

Note that some links returned will not in fact lead you to full text
content as you would understandbly think and expect. That is, if you
use the \code{filter} parameter with e.g., \code{\link[rcrossref:cr_works]{rcrossref::cr_works()}}
and filter to only full text content, some links may actually give back
only metadata for an article. Elsevier is perhaps the worst offender,
for one because they have a lot of entries in Crossref TDM, but most
of the links that are apparently full text are not in facct full text,
but only metadata.

Check out \code{\link[=auth]{auth()}} for details on authentication.
}
\examples{
\dontrun{
## pensoft
data(dois_pensoft)
(links <- crm_links(dois_pensoft[1], "all"))
### xml
crm_text(url=links, type='xml')
### pdf
crm_text(url=links, type="pdf", read = FALSE)
crm_text(links, "pdf")

## hindawi
data(dois_pensoft)
(links <- crm_links(dois_pensoft[1], "all"))
### xml
crm_text(links, 'xml')
### pdf
crm_text(links, "pdf", read=FALSE)
crm_text(links, "pdf")

## DOIs w/ full text, and with CC-BY 3.0 license
data(dois_crminer_ccby3)
(links <- crm_links(dois_crminer_ccby3[40], "all"))
# crm_text(links, 'xml')

## You can use crm_xml, crm_plain, and crm_pdf to go directly to
## that format
(links <- crm_links(dois_crminer_ccby3[5], "all"))
crm_xml(links)
#crm_pdf(links)

### Caching, for PDFs
# out <- cr_members(2258, filter=c(has_full_text = TRUE), works = TRUE)
# (links <- crm_links(out$data$DOI[10], "all"))
# crm_text(links, type = "pdf", cache=FALSE)
# system.time( cacheyes <- crm_text(links, type = "pdf", cache=TRUE) )
### second time should be faster
# system.time( cacheyes <- crm_text(links, type = "pdf", cache=TRUE) )
# system.time( cacheno <- crm_text(links, type = "pdf", cache=FALSE) )
# identical(cacheyes, cacheno)

## elsevier
## requires extra authentication
data(dois_elsevier)

## set key first
# Sys.setenv(CROSSREF_TDM_ELSEVIER = "your-key")
## XML
link <- crm_links("10.1016/j.funeco.2010.11.003", "xml")
# res <- crm_text(url = link, type = "xml")
## plain text
link <- crm_links("10.1016/j.funeco.2010.11.003", "plain")
# res <- crm_text(url = link, "plain")

## Wiley
Sys.setenv(CROSSREF_TDM = "your-key")

### all wiley
data(dois_wiley)

res <- list()
dois <- dois_wiley$set1[1:10]
for (i in seq_along(dois)) {
  tmp <- crm_links(dois[i], "all")
  res[[i]] <- crm_text(url = tmp, type = "pdf", cache=FALSE,
      overwrite_unspecified = TRUE)
}
res

#### older dates
# res <- list()
# for (i in seq_along(dois_wiley$set2)) {
#   tmp <- crm_links(dois_wiley$set2[i], "all")
#   res[[i]] <- crm_text(tmp, type = "pdf", cache=FALSE,
#         overwrite_unspecified=TRUE)
# }
# res

### wiley subset with CC By 4.0 license
# res <- list()
# for (i in seq_along(dois_wiley$set3)) {
#   tmp <- crm_links(dois_wiley$set3[i], "all")
#   res[[i]] <- crm_text(tmp, type = "pdf", cache=F,
        overwrite_unspecified=TRUE)
# }
}
}
