% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crminer-package.R
\docType{package}
\name{crminer-package}
\alias{crminer}
\alias{crminer-package}
\title{crminer - Crossref text miner}
\description{
crminer - Crossref text miner
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

