% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crminer-package.R
\docType{data}
\name{dois_crminer}
\alias{dois_crminer}
\title{A character vector of 500 DOIs from Crossref}
\format{A character vector of length 500}
\description{
Obtained via
\code{rcrossref::cr_works(filter = c(has_full_text = TRUE), limit = 500)}
}
\keyword{datasets}
