% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crminer-package.R
\docType{data}
\name{dois_hindawi}
\alias{dois_hindawi}
\title{A character vector of 50 Hindawi DOIs from Crossref}
\format{A character vector of length 50}
\description{
Obtained via \code{rcrossref::cr_members(98, filter = c(has_full_text = TRUE), works = TRUE, limit = 50)}
}
\keyword{datasets}
