% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crminer-package.R
\docType{data}
\name{dois_pensoft}
\alias{dois_pensoft}
\title{A character vector of 100 Pensoft DOIs from Crossref}
\format{A character vector of length 100}
\description{
Obtained via \code{rcrossref::cr_members(2258, filter = c(has_full_text = TRUE), works = TRUE, limit = 100)}
}
\keyword{datasets}
