% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_extract.R
\name{crm_extract}
\alias{crm_extract}
\title{Extract text from a single pdf document}
\usage{
crm_extract(path, ...)
}
\arguments{
\item{path}{(character) Path to a file. required}

\item{...}{args passed on to \code{\link[pdftools:pdf_info]{pdftools::pdf_info()}}
and \code{\link[pdftools:pdf_text]{pdftools::pdf_text()}} - any args are passed to
both of those function calls, which makes sense}
}
\value{
An object of class \code{crm_pdf} with a slot for
\code{info} (pdf metadata essentially), and \code{text} (the extracted
text) - with an attribute (\code{path}) with the path to the pdf
on disk
}
\description{
Extract text from a single pdf document
}
\details{
We use \pkg{pdftools} under the hood to do pdf text
extraction
}
\examples{
path <- system.file("examples", "MairChamberlain2014RJournal.pdf",
   package = "crminer")
(res <- crm_extract(path))
res$info
res$text
# with newlines, pretty print
cat(res$text)

# another example
path <- system.file("examples", "ChamberlainEtal2013Ecosphere.pdf",
   package = "crminer")
(res <- crm_extract(path))
res$info
cat(res$text)
}
