% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crminer-package.R
\docType{package}
\name{crminer-package}
\alias{crminer-package}
\alias{crminer}
\title{crminer - Crossref text miner}
\description{
crminer - Crossref text miner
}
\section{Package API (functions)}{

\itemize{
\item \code{\link[=crm_links]{crm_links()}} - get full text links from DOIs
\item \code{\link[=as_tdmurl]{as_tdmurl()}} - coerce a URL to a tdmurl
\item \code{\link[=crm_text]{crm_text()}} - general purpose interface to request
full text
\item \code{\link[=crm_pdf]{crm_pdf()}} - request pdf full text
\item \code{\link[=crm_plain]{crm_plain()}} - request plain text full text
\item \code{\link[=crm_xml]{crm_xml()}} - request xml full text
\item \code{\link[=crm_extract]{crm_extract()}} - extract text from a pdf
}
}

\section{Authentication}{

You should first start reading up on authentication (\code{\link[=auth]{auth()}})
since you are probably here to do text mining, and most of the full text
links available via Crossref are behind authentication.
}

\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
